Separating Printer from CIPrinter.
