+++ cibyl/models/ci/printers/__init__.py
@@ -14,27 +14,14 @@
 #    under the License.
 """
 from abc import ABC, abstractmethod
-from enum import Enum
+
+from cibyl.cli.output import OutputStyle
+from cibyl.models.ci.printers.colored import ColoredPrinter
+from cibyl.models.ci.printers.raw import RawPrinter
+from cibyl.publisher import Printer
 
 
-class PrintMode(Enum):
-    SIMPLE = 0
-    COMPLETE = 1
-
-
-class Printer(ABC):
-    def __init__(self, mode=PrintMode.COMPLETE, verbosity=0):
-        self._mode = mode
-        self._verbosity = verbosity
-
-    @property
-    def mode(self):
-        return self._mode
-
-    @property
-    def verbosity(self):
-        return self._verbosity
-
+class CIPrinter(ABC, Printer):
     @abstractmethod
     def print_environment(self, env):
         raise NotImplementedError
@@ -54,3 +41,27 @@ class Printer(ABC):
     @abstractmethod
     def print_test(self, test):
         raise NotImplementedError
+
+
+class CIPrinterFactory:
+    @staticmethod
+    def from_style(style, mode, verbosity):
+        """
+
+        :param style:
+        :type style: :class:`OutputStyle`
+        :param mode:
+        :type mode: :class:`PrintMode`
+        :param verbosity:
+        :type verbosity: int
+        :return:
+        :rtype: :class:`CIPrinter`
+        :raise NotImplementedError: If there is not printer for the desired
+            style.
+        """
+        if style == OutputStyle.TEXT:
+            return RawPrinter(mode, verbosity)
+        elif style == OutputStyle.COLORIZED:
+            return ColoredPrinter(mode, verbosity)
+        else:
+            raise NotImplementedError(f'Unknown output style: {style}')
